+++ packages/adapter-react/src/index.ts
@@ -1,3 +1,4 @@
 export { useValfuseForm } from "./use-valfuse-form";
+export type { UseValfuseFormReturn, ValfuseFieldError } from "./use-valfuse-form";
 export { createValfuseResolver } from "./create-valfuse-resolver";